Much better uninstall script.
For each file installed with install.sh, the file will always be
restored from the backup file if it exists. If not, then an ls -lh on
the file in question will be displayed, after which the user will asked
to decide whether to delete the file. If the -y flag is passed to the
script, then files without backups will be deleted without confirmation.