Use find to locate wallpaper image.
Rather than using shell expansion to try to find a wallpaper image in
$HOME, use find to avoid generating an error message when no wallpaper
image is found (zsh warns when shell expansions are empty). Crucially,
set find's max depth to 1 to avoid recursively searching all of $HOME,
which can add a lot of time to opening up a new shell session.