Recursively update submodules
Some vim plugins, which are all installed as submodules, use submodules
themselves, so the `--recursive` flag needs to be here for those to be
properly installed.

Additionally, git is now smart enough to only initialize the submodules
that need to be initialized when the `--init` flag is passed to `git
submodule update`, so all the submodule setup can now happen in one git
command.