Add latest paredit from vimscripts, fix vimrc
The '| edit!' at the end of the write sudo shortcut broke paredit for
some reason. Fortunately, removing it doesn't seem to impact that
shortuct.

Increase the leader timeout to 750 ms because ';O' is a common paredit
move, and difficult on a programmer dvorak keyboard that's missing the
cap on the 'o' key.